fix(clerk-js): Remove deeply cloned check
